% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{absolute_residual_scores}
\alias{absolute_residual_scores}
\title{Absolute Residual Scores}
\usage{
absolute_residual_scores(y_pred, y_true)
}
\arguments{
\item{y_pred}{Numeric vector or matrix. The predictions made by the model.}

\item{y_true}{Numeric vector or matrix. The true target values.}
}
\value{
Numeric vector or matrix. Residual normalized scores for each prediction.
}
\description{
Calculate nonconformity scores based on absolute differences between the predicted and true values.
}
