% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithmique-package.R
\docType{package}
\name{Algorithmique-package}
\alias{Algorithmique}
\alias{Algorithmique-package}
\title{Algorithmique: Conformal Prediction in High Dimensions}
\description{
To add.
}
\author{
\strong{Maintainer}: Fatima Zahra HANNOU \email{fatimzahrahannou@gmail.com}

Authors:
\itemize{
  \item Jaad BELHOUARI \email{jaad.belhouari@gmail.com}
  \item Ayman MIMOUN \email{aym.mimoun@gmail.com}
}

}
\keyword{internal}
