% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{column_wise_rank}
\alias{column_wise_rank}
\title{Column-wise Rank}
\usage{
column_wise_rank(matrix)
}
\arguments{
\item{matrix}{A numeric matrix of size n x m.}
}
\value{
A matrix of size n x m containing column-wise ranks.
}
\description{
Returns a matrix of the same size where each element represents
the column-wise rank of the corresponding element in the input matrix.
}
