% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_and_save_dataset_vectorized}
\alias{generate_and_save_dataset_vectorized}
\title{Generate and Save a Non-linear Dataset (Optimized)}
\usage{
generate_and_save_dataset_vectorized(Y_dim, n_samples, name_dataset_file)
}
\arguments{
\item{Y_dim}{Integer. The number of output dimensions for Y.}

\item{n_samples}{Integer. Number of samples to generate (e.g. 1e6).}

\item{name_dataset_file}{String. Name of the output CSV file.}
}
\value{
None. Writes the dataset to a CSV file.
}
\description{
This function generates a synthetic dataset with two input features (X1, X2) and a multi-dimensional
target variable Y using vectorized operations. The dataset is saved as a CSV file.
}
\examples{
generate_and_save_dataset_vectorized(24, 1e6, "dataset.csv")

}
