% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R, R/uncertainty_utils.R
\name{get_Beta_quantiles}
\alias{get_Beta_quantiles}
\title{Beta Quantile Calculation}
\usage{
get_Beta_quantiles(sorted_Scores, Beta)

get_Beta_quantiles(sorted_Scores, Beta)
}
\arguments{
\item{sorted_Scores}{Numeric matrix. The sorted score matrix (from lowest to highest score values).}

\item{Beta}{Numeric value. The error rate to determine the quantile.}
}
\value{
Numeric vector. Beta score quantiles.

Numeric vector. The Beta quantiles for each target dimension.
}
\description{
Get the Beta score quantile for each dimension.

Calculate the Beta quantiles from the sorted score matrix for each target dimension.
}
