% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{get_coverage_and_intWidth}
\alias{get_coverage_and_intWidth}
\title{Coverage and Interval Width Calculation}
\usage{
get_coverage_and_intWidth(y_true, y_low, y_up)
}
\arguments{
\item{y_true}{Numeric matrix. True target values.}

\item{y_low}{Numeric matrix. Lower prediction bounds.}

\item{y_up}{Numeric matrix. Upper prediction bounds.}
}
\value{
List containing:
\item{coverage}{Numeric vector. Coverage per target dimension.}
\item{interval_width}{Numeric vector. Average interval width per target dimension.}
}
\description{
Calculate the empirical coverage and the average interval width for each target dimension.
}
