% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{get_r_max_scalar}
\alias{get_r_max_scalar}
\title{Compute r_max Scalar}
\usage{
get_r_max_scalar(R_max, alpha)
}
\arguments{
\item{R_max}{A numeric vector containing max ranks for each observation.}

\item{alpha}{A numeric value (between 0 and 1) representing the error rate.}
}
\value{
A scalar integer corresponding to the quantile threshold.
}
\description{
Gets the scalar value corresponding to the (1 - alpha) quantile
of the R_max vector.
}
