% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R, R/uncertainty_utils.R
\name{get_sorted_scores}
\alias{get_sorted_scores}
\title{Sorted Scores Calculation}
\usage{
get_sorted_scores(Scores)

get_sorted_scores(Scores)
}
\arguments{
\item{Scores}{Numeric matrix. The score matrix to be sorted.}
}
\value{
Numeric matrix. Sorted scores.

Numeric matrix. The sorted scores for each target dimension.
}
\description{
Sort the score for each dimension.

Sort the score matrix for each target dimension in ascending order.
}
