% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training_GB.R
\name{MLModel}
\alias{MLModel}
\title{MLModel}
\format{
An \code{R6Class} object with the following public methods:
\describe{
\item{\code{new(X, Y, method = "polynomial", degree = 2, ...)}}{Constructor for the model. Initializes data and method.}
\item{\code{fit()}}{Trains one model per output dimension in \code{Y}.}
\item{\code{predict(X_new)}}{Predicts all output values from new input data \code{X_new}.}
}
}
\description{
R6 class for multi-output regression using either polynomial regression or gradient boosting with bagging (via xgboost).
}
\examples{
X <- X_train
Y <- Y_train
model <- MLModel$new(X, Y, method = "polynomial")
model$fit()
Y_pred <- model$predict(X)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLModel-new}{\code{MLModel$new()}}
\item \href{#method-MLModel-fit}{\code{MLModel$fit()}}
\item \href{#method-MLModel-predict}{\code{MLModel$predict()}}
\item \href{#method-MLModel-clone}{\code{MLModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$new(
  X,
  Y,
  method = "polynomial",
  degree = 2,
  n_estimators = 10,
  nrounds = 100,
  subsample_rate = 0.8
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$predict(X_new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
