% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelUncertainties.R
\name{ModelUncertainties}
\alias{ModelUncertainties}
\title{ModelUncertainties Class}
\description{
This class encapsulates the process of fitting uncertainty models and making predictions with uncertainty intervals.
The class supports multiple uncertainty estimation methods such as "Beta_Optim" and "Max_Rank".
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The machine learning model (e.g., regression, classifier).}

\item{\code{X_calibration}}{Calibration dataset features.}

\item{\code{Y_calibration}}{Calibration dataset targets.}

\item{\code{uncertainty_method}}{The method for uncertainty estimation (e.g., "Beta_Optim").}

\item{\code{Global_alpha}}{The significance level for uncertainty estimation.}

\item{\code{Beta_optim}}{The optimized Beta value used for uncertainty estimation.}

\item{\code{quantiles}}{The quantiles used for uncertainty intervals.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelUncertainties-new}{\code{ModelUncertainties$new()}}
\item \href{#method-ModelUncertainties-fit}{\code{ModelUncertainties$fit()}}
\item \href{#method-ModelUncertainties-predict}{\code{ModelUncertainties$predict()}}
\item \href{#method-ModelUncertainties-clone}{\code{ModelUncertainties$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUncertainties-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUncertainties-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the ModelUncertainties class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUncertainties$new(
  model,
  X_calibration,
  Y_calibration,
  uncertainty_method,
  Global_alpha
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The machine learning model (e.g., regression, classifier).}

\item{\code{X_calibration}}{Calibration dataset features.}

\item{\code{Y_calibration}}{Calibration dataset targets.}

\item{\code{uncertainty_method}}{The method for uncertainty estimation (e.g., "Beta_Optim").}

\item{\code{Global_alpha}}{The significance level for uncertainty estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUncertainties-fit"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUncertainties-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the uncertainty model based on the chosen method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUncertainties$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUncertainties-predict"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUncertainties-predict}{}}}
\subsection{Method \code{predict()}}{
Predicts uncertainty bounds for the given test data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUncertainties$predict(X_test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{Features of the test dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with two elements: \code{y_lower} and \code{y_upper}, representing the lower and upper bounds.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelUncertainties-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelUncertainties-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelUncertainties$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
