% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_split.R
\name{train_test_calib_split}
\alias{train_test_calib_split}
\title{Split data into training, calibration, and test sets (with indices)}
\usage{
train_test_calib_split(X, Y, n_train = 1000, n_test = 1000, n_calib)
}
\arguments{
\item{X}{Input features}

\item{Y}{Output targets}

\item{n_train}{Number of training samples (default 1000)}

\item{n_test}{Number of test samples (default 1000)}

\item{n_calib}{Number of calibration samples (user-defined)}
}
\value{
A list: data splits + row indices
}
\description{
Split data into training, calibration, and test sets (with indices)
}
