% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{get_Beta_quantiles}
\alias{get_Beta_quantiles}
\title{Get Beta Quantiles (Vectorized)}
\usage{
get_Beta_quantiles(sorted_Scores, Beta)
}
\arguments{
\item{sorted_Scores}{Numeric matrix. The sorted score matrix (from lowest to highest score values).}

\item{Beta}{Numeric value. The error rate to determine the quantile.}
}
\value{
Numeric vector. The Beta quantiles for each target dimension.
}
\description{
Calculate the Beta quantiles from the sorted score matrix for each target dimension.
}
