% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{get_prediction_bounds}
\alias{get_prediction_bounds}
\title{Get Prediction Bounds}
\usage{
get_prediction_bounds(quantile, y_pred)
}
\arguments{
\item{quantile}{Numeric vector. The Beta quantiles.}

\item{y_pred}{Numeric matrix. The predicted values for each target dimension.}
}
\value{
A list containing:
\item{y_lower}{Numeric matrix. The lower prediction bounds.}
\item{y_upper}{Numeric matrix. The upper prediction bounds.}
}
\description{
Calculate the lower and upper prediction bounds given the Beta quantiles and predicted values.
}
