% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{get_sorted_scores}
\alias{get_sorted_scores}
\title{Get Sorted Scores}
\usage{
get_sorted_scores(Scores)
}
\arguments{
\item{Scores}{Numeric matrix. The score matrix to be sorted.}
}
\value{
Numeric matrix. The sorted scores for each target dimension.
}
\description{
Sort the score matrix for each target dimension in ascending order.
}
