% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_utils.R
\name{simultaneous_coverage}
\alias{simultaneous_coverage}
\title{Simultaneous Coverage Calculation}
\usage{
simultaneous_coverage(y_true, y_lower, y_upper)
}
\arguments{
\item{y_true}{Numeric matrix. The true target values.}

\item{y_lower}{Numeric matrix. The lower bound of the prediction interval.}

\item{y_upper}{Numeric matrix. The upper bound of the prediction interval.}
}
\value{
Numeric value. The empirical simultaneous coverage.
}
\description{
Calculate the empirical simultaneous coverage, which measures the percentage of cases where all true values
fall within their respective prediction intervals.
}
